Rename ExternalDevice to ExternalLab; Rename ExternalDeviceCommand to ExternalLabCommand; Rename ExternalLabCollection to ExternalLabsCollection; Rename ExternalServiceRequest to ExternalLabsServiceInformation
