+++ DMX.Core.Api/Models/External/ExternalLabs/ExternalLabCommand.cs
@@ -2,11 +2,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ---------------------------------------------------------------
 
+using System.Collections.Generic;
+
 namespace DMX.Core.Api.Models.External.ExternalLabs
 {
-    public class ExternalLabServiceRequest
+    public class ExternalLabCommand
     {
-        public string ServiceType { get; set; }
-        public string ServiceId { get; set; }
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public IDictionary<string, string> Parameters { get; set; }
     }
 }
