+++ DMX.Core.Api/Models/External/ExternalLabs/ExternalLabsCollection.cs
@@ -2,14 +2,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ---------------------------------------------------------------
 
-using DMX.Core.Api.Models.Devices.External.ExternalDevices;
 using Newtonsoft.Json;
 
 namespace DMX.Core.Api.Models.External.ExternalLabs
 {
-    public class ExternalLabCollection
+    public class ExternalLabsCollection
     {
         [JsonProperty("Data")]
-        public ExternalDevice[] Devices { get; set; }
+        public ExternalLab[] Devices { get; set; }
     }
 }
