+++ DMX.Core.Api/Models/External/ExternalLabs/ExternalLabsServiceInformation.cs
@@ -1,0 +1,12 @@
+﻿// ---------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// ---------------------------------------------------------------
+
+namespace DMX.Core.Api.Models.External.ExternalLabs
+{
+    public class ExternalLabsServiceInformation
+    {
+        public string ServiceType { get; set; }
+        public string ServiceId { get; set; }
+    }
+}
